refactor 로그인시 RefreshToken 중복저장 이슈 수정

+++ src/main/java/com/sparta/gimbap_heaven/user/Entity/User.java
@@ -37,7 +37,7 @@ public class User {
     private String intro;
 
     @Column
-    private Integer money;
+    private double money;
 
     public User(String username, String password, String email, UserRoleEnum role) {
         this.username = username;
